refactor [Delivers #170166651]
